moves all expensecomponents to single Expense folder
